Dev

See merge request via-profit-services/file-storage!19

+++ src/schemas/file-storage/resolvers/DeleteFiles.ts
@@ -1,4 +1,4 @@
-import { ServerError } from '@via-profit-services/core';
+import { ServerError, TWhereAction } from '@via-profit-services/core';
 import { IFieldResolver } from 'graphql-tools';
 
 import createLoaders from '../loaders';
@@ -6,13 +6,14 @@ import FileStorageService from '../service';
 import { ExtendedContext } from '../types';
 
 interface TArgs {
-  ids: string[];
+  ids?: string[];
+  owners?: string[];
 }
 
 const DeleteFilesResolver: IFieldResolver<any, ExtendedContext, TArgs> = async (
   parent, args, context,
 ) => {
-  const { ids } = args;
+  const { ids, owners } = args;
   const { logger, token } = context;
   const { uuid } = token;
 
@@ -20,24 +21,41 @@ const DeleteFilesResolver: IFieldResolver<any, ExtendedContext, TArgs> = async (
   const loaders = createLoaders(context);
 
   let deletedIds: string[] = [];
-  try {
-    deletedIds = await fileService.deleteStaticFiles(ids);
-  } catch (err) {
-    logger.fileStorage.error('Failed to Delete files', { err, uuid });
-    throw new ServerError('Failed to Delete files', { err, uuid });
+
+  // get by IDs
+  if (ids && Array.isArray(ids)) {
+    try {
+      logger.fileStorage.debug(`Will be deleted ${ids.length} file(s)`, { uuid });
+      deletedIds = await fileService.deleteStaticFiles(ids);
+    } catch (err) {
+      logger.fileStorage.error('Failed to Delete files', { err, uuid });
+      throw new ServerError('Failed to Delete files', { err, uuid });
+    }
   }
 
+  if (owners && Array.isArray(owners)) {
+    const files = await fileService.getFiles({
+      where: [
+        ['owner', TWhereAction.IN, owners],
+      ],
+    });
+
+    const idsByOwner = files.nodes.map((node) => node.id);
+    if (idsByOwner.length) {
+      try {
+        logger.fileStorage.debug(`Will be deleted ${idsByOwner} file(s)`, { uuid });
+        deletedIds = await fileService.deleteStaticFiles(idsByOwner);
+      } catch (err) {
+        logger.fileStorage.error('Failed to Delete files', { err, uuid });
+        throw new ServerError('Failed to Delete files', { err, uuid });
+      }
+    }
+  }
 
   deletedIds.forEach((id) => {
     loaders.files.clear(id);
     logger.fileStorage.debug(`File ${id} was deleted. Initiator: Account ${uuid}`);
   });
-
-  if (deletedIds.length !== ids.length) {
-    logger.fileStorage.debug(
-      'Not all files were deleted, because some of them were not found in the database',
-    );
-  }
 
   return true;
 };
